feat(vscode-webui): extract CreatePrAction component and use it in AttemptCompletionTool and AttemptTodoCompletionView
This refactor extracts the "Create PR" button logic into a reusable `CreatePrAction` component. It is now used in both the standard `AttemptCompletionTool` and the `AttemptTodoCompletionView` (the completion view for `newTask` sub-agents), ensuring a consistent experience for creating pull requests across different task types.

- Extracted `CreatePrAction` to `packages/vscode-webui/src/features/tools/components/create-pr-action.tsx`
- Replaced inline PR button logic in `AttemptCompletionTool` with `CreatePrAction`
- Added `CreatePrAction` to `AttemptTodoCompletionView` header actions
